+++ .github/workflows/pages.yml
@@ -77,6 +77,7 @@ jobs:
       - name: Build with Gatsby
         env:
           PREFIX_PATHS: 'true'
+          GA_TACKING_ID: 'G-YTHDX2HH7G'
         run: ${{ steps.detect-package-manager.outputs.manager }} run build
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v1
