+++ src/utils/typography.js
@@ -26,3 +26,5 @@ if (process.env.NODE_ENV !== `production`) {
 export default typography
 export const rhythm = typography.rhythm
 export const scale = typography.scale
+
+export const BLOG_PATH = "/decompile"
